+++ src/NonlinearEigenproblems.jl
@@ -18,5 +18,7 @@ include(joinpath("utils", "Serialization.jl"))
 include(joinpath("rk_helper", "RKHelper.jl"))
 @include_export("NEPSolver.jl", NEPSolver)
 @include_export("Gallery.jl", Gallery)
+@include_export("NEPTransformations.jl", NEPTransformations)
+
 
 end
