+++ src/nleigs_coefficients.jl
@@ -3,6 +3,7 @@ using Random
 using SparseArrays
 using IterativeSolvers
 using NonlinearEigenproblems
+using NonlinearEigenproblems.RKHelper
 
 export nleigs_coefficients
 
@@ -68,12 +69,12 @@ function nleigs_coefficients(
         return_details::Bool = false
         ) where {T<:Real, CT<:Complex{T}}
 
-#    @parse_logger_param!(logger)
+    @parse_logger_param!(logger)
 
     # The following variables are used when creating the return values, so put them in scope
     D = Vector{Matrix{CT}}()
 
-    P = NonlinearEigenproblems.RKHelper.get_rk_nep(T, nep)
+    P = get_rk_nep(T, nep)
     n = size(nep, 1)
     n == 1 && (maxdgr = maxit + 1)
     computeD = (n <= 400) # for small problems, explicitly use generalized divided differences
@@ -87,22 +88,22 @@ function nleigs_coefficients(
         if isempty(nodes)
             error("Interpolation nodes must be provided via 'nodes' when no Leja-Bagby points ('leja' == 0) are used.")
         end
-        gamma,_ = NonlinearEigenproblems.RKHelper.discretizepolygon(Σ)
+        gamma,_ = discretizepolygon(Σ)
         max_count = max(maxit,maxdgr)+2
         σ = repeat(reshape(nodes, :, 1), ceil(Int, max_count/length(nodes)), 1)
-        _,ξ,β = NonlinearEigenproblems.RKHelper.lejabagby(σ[1:maxdgr+2], Ξ, gamma, maxdgr+2, true, P.p)
+        _,ξ,β = lejabagby(σ[1:maxdgr+2], Ξ, gamma, maxdgr+2, true, P.p)
     elseif leja == 1 # use leja nodes in expansion phase
         if isempty(nodes)
-            gamma,nodes = NonlinearEigenproblems.RKHelper.discretizepolygon(Σ, true)
+            gamma,nodes = discretizepolygon(Σ, true)
         else
-            gamma,_ = NonlinearEigenproblems.RKHelper.discretizepolygon(Σ)
+            gamma,_ = discretizepolygon(Σ)
         end
         nodes = repeat(reshape(nodes, :, 1), ceil(Int, (maxit+1)/length(nodes)), 1)
-        σ,ξ,β = NonlinearEigenproblems.RKHelper.lejabagby(gamma, Ξ, gamma, maxdgr+2, false, P.p)
+        σ,ξ,β = lejabagby(gamma, Ξ, gamma, maxdgr+2, false, P.p)
     else # use leja nodes in both phases
-        gamma,_ = NonlinearEigenproblems.RKHelper.discretizepolygon(Σ)
+        gamma,_ = discretizepolygon(Σ)
         max_count =  max(maxit,maxdgr)+2
-        σ,ξ,β = NonlinearEigenproblems.RKHelper.lejabagby(gamma, Ξ, gamma, max_count, false, P.p)
+        σ,ξ,β = lejabagby(gamma, Ξ, gamma, max_count, false, P.p)
     end
     ξ[maxdgr+2] = NaN # not used
     if (!P.spmf || !isfunm) && length(σ) != length(unique(σ))
@@ -118,7 +119,7 @@ function nleigs_coefficients(
         sgdd = Matrix{CT}(undef, 0, 0)
     else
         # Compute scalar generalized divided differences
-        sgdd = NonlinearEigenproblems.RKHelper.scgendivdiffs(σ[range], ξ[range], β[range], maxdgr, isfunm, get_fv(nep))
+        sgdd = RKHelper.scgendivdiffs(σ[range], ξ[range], β[range], maxdgr, isfunm, get_fv(nep))
         # Construct first generalized divided difference
         computeD && push!(D, constructD(0, P, sgdd))
         # Norm of first generalized divided difference
@@ -178,10 +179,10 @@ function nleigs_coefficients(
                     nrmD = nrmD[1:k]
 
                     N -= 1
-                    #NonlinearEigenproblems.push_info!(logger,
-                    #           "Linearization converged after $kconv iterations")
-                    #NonlinearEigenproblems.push_info!(logger,
-                    #           "--> freeze linearization")
+                    push_info!(logger,
+                               "Linearization converged after $kconv iterations")
+                    push_info!(logger,
+                               "--> freeze linearization")
                 elseif k == maxdgr+1
                     kconv = k
                     expand = false
@@ -194,8 +195,8 @@ function nleigs_coefficients(
 
                     N -= 1
                     @warn "NLEIGS: Linearization not converged after $maxdgr iterations"
-                    #NonlinearEigenproblems.push_info!(logger,
-                    #           "--> freeze linearization")
+                    push_info!(logger,
+                               "--> freeze linearization")
                 end
             end
         end
